dynamique + transition tsm et csm actives

+++ src/Train_dynamique.cpp
@@ -1,7 +1,7 @@
 #include "Train_dynamique.hpp"
 
-int Train_dynamique::getV_train(){return V_train;}
-void Train_dynamique::setV_train(int V){V_train = V;}
+float Train_dynamique::getV_train(){return V_train;}
+void Train_dynamique::setV_train(float V){V_train = V;}
 
 /*void Train_dynamique::TD_update()
 {
