+++ src/Train_dynamique.hpp
@@ -6,7 +6,7 @@
 class Train_dynamique
 {
 	private :
-	int V_train = 80;
+	float V_train = 80;
 	//Clock chrono;//créer le chrono
 	//Time diftime;//créer la varaible qui stocke le temps écoulé
 	//float deltats;//créer la variable qui stocke le temps écoulé en seconde*/
@@ -14,9 +14,8 @@ class Train_dynamique
 	public :
 
 	//void TD_update();
-	int getV_train();
-	void setV_train(int V);
-	//void mouvementTrain();
+	float getV_train();
+	void setV_train(float V);
 
 };
